+++ src/com/amplitude/api/PinnedAmplitudeClient.java
@@ -1,6 +1,7 @@
 package com.amplitude.api;
 
 import android.content.Context;
+import android.text.TextUtils;
 
 import java.io.IOException;
 import java.security.GeneralSecurityException;
@@ -8,7 +9,9 @@ import java.security.KeyStore;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSocketFactory;
@@ -135,16 +138,35 @@ public class PinnedAmplitudeClient extends AmplitudeClient {
         }
     }
 
-    protected static PinnedAmplitudeClient instance = new PinnedAmplitudeClient();
+    static Map<String, PinnedAmplitudeClient> instances =
+        new HashMap<String, PinnedAmplitudeClient>();
 
     public static PinnedAmplitudeClient getInstance() {
-        return instance;
+        return getInstance(null);
+    }
+
+    public static synchronized PinnedAmplitudeClient getInstance(String instance) {
+        if (TextUtils.isEmpty(instance)) {
+            instance = Constants.DEFAULT_INSTANCE;
+        }
+        instance = instance.toLowerCase();
+
+        PinnedAmplitudeClient client = instances.get(instance);
+        if (client == null) {
+            client = new PinnedAmplitudeClient(instance);
+            instances.put(instance, client);
+        }
+        return client;
     }
 
     protected SSLSocketFactory sslSocketFactory;
 
     public PinnedAmplitudeClient() {
         super();
+    }
+
+    public PinnedAmplitudeClient(String instance) {
+        super(instance);
     }
 
     protected boolean initializedSSLSocketFactory = false;