+++ src/com/amplitude/api/AmplitudeClient.java
@@ -9,11 +9,6 @@ import android.text.TextUtils;
 import android.util.Pair;
 
 import com.amplitude.security.MD5;
-import com.squareup.okhttp.FormEncodingBuilder;
-import com.squareup.okhttp.OkHttpClient;
-import com.squareup.okhttp.Request;
-import com.squareup.okhttp.RequestBody;
-import com.squareup.okhttp.Response;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -28,6 +23,11 @@ import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicBoolean;
+
+import okhttp3.FormBody;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
 
 public class AmplitudeClient {
 
@@ -45,13 +45,14 @@ public class AmplitudeClient {
         return instance;
     }
 
-    private static AmplitudeLog logger = AmplitudeLog.getLogger();
+    private static final AmplitudeLog logger = AmplitudeLog.getLogger();
 
     protected Context context;
     protected OkHttpClient httpClient;
     protected String apiKey;
     protected String userId;
     protected String deviceId;
+    protected String sharedPreferencesName;
     private boolean newDeviceIdPerInstall = false;
     private boolean useAdvertisingIdForDeviceId = false;
     private boolean initialized = false;
@@ -108,9 +109,10 @@ public class AmplitudeClient {
             this.context = context.getApplicationContext();
             this.httpClient = new OkHttpClient();
             this.apiKey = apiKey;
+            this.sharedPreferencesName = getSharedPreferencesName();
             initializeDeviceInfo();
             SharedPreferences preferences = context.getSharedPreferences(
-                    getSharedPreferencesName(), Context.MODE_PRIVATE);
+                    this.sharedPreferencesName, Context.MODE_PRIVATE);
             if (userId != null) {
                 this.userId = userId;
                 preferences.edit().putString(Constants.PREFKEY_USER_ID, userId).commit();
@@ -222,7 +224,7 @@ public class AmplitudeClient {
         this.optOut = optOut;
 
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         preferences.edit().putBoolean(Constants.PREFKEY_OPT_OUT, optOut).commit();
         return instance;
     }
@@ -440,49 +442,49 @@ public class AmplitudeClient {
 
     long getLastEventTime() {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         return preferences.getLong(Constants.PREFKEY_LAST_EVENT_TIME, -1);
     }
 
     void setLastEventTime(long timestamp) {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         preferences.edit().putLong(Constants.PREFKEY_LAST_EVENT_TIME, timestamp).commit();
     }
 
     long getLastEventId() {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         return preferences.getLong(Constants.PREFKEY_LAST_EVENT_ID, -1);
     }
 
     void setLastEventId(long eventId) {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         preferences.edit().putLong(Constants.PREFKEY_LAST_EVENT_ID, eventId).commit();
     }
 
     long getLastIdentifyId() {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         return preferences.getLong(Constants.PREFKEY_LAST_IDENTIFY_ID, -1);
     }
 
     void setLastIdentifyId(long identifyId) {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         preferences.edit().putLong(Constants.PREFKEY_LAST_IDENTIFY_ID, identifyId).commit();
     }
 
     long getPreviousSessionId() {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         return preferences.getLong(Constants.PREFKEY_PREVIOUS_SESSION_ID, -1);
     }
 
     void setPreviousSessionId(long timestamp) {
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         preferences.edit().putLong(Constants.PREFKEY_PREVIOUS_SESSION_ID, timestamp).commit();
     }
 
@@ -741,7 +743,7 @@ public class AmplitudeClient {
 
         this.userId = userId;
         SharedPreferences preferences = context.getSharedPreferences(
-                getSharedPreferencesName(), Context.MODE_PRIVATE);
+                sharedPreferencesName, Context.MODE_PRIVATE);
         preferences.edit().putString(Constants.PREFKEY_USER_ID, userId).commit();
         return instance;
     }
@@ -754,7 +756,8 @@ public class AmplitudeClient {
         }
 
         this.deviceId = deviceId;
-        DatabaseHelper.getDatabaseHelper(context).insertOrReplaceKeyValue(DEVICE_ID_KEY, deviceId);
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        dbHelper.insertOrReplaceKeyValue(DEVICE_ID_KEY, deviceId);
         return instance;
     }
 
@@ -892,7 +895,7 @@ public class AmplitudeClient {
             logger.e(TAG, e.toString());
         }
 
-        RequestBody body = new FormEncodingBuilder()
+        FormBody body = new FormBody.Builder()
             .add("v", apiVersionString)
             .add("client", apiKey)
             .add("e", events)
@@ -1218,16 +1221,20 @@ public class AmplitudeClient {
             sourcePkgName = Constants.PACKAGE_NAME;
         }
 
+        // only perform upgrade if deviceId not yet in database
+        DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
+        if (!TextUtils.isEmpty(dbHelper.getValue(DEVICE_ID_KEY))) {
+            return true;
+        }
+
         String prefsName = sourcePkgName + "." + context.getPackageName();
         SharedPreferences preferences =
                 context.getSharedPreferences(prefsName, Context.MODE_PRIVATE);
-
         String deviceId = preferences.getString(Constants.PREFKEY_DEVICE_ID, null);
         if (!TextUtils.isEmpty(deviceId)) {
-            DatabaseHelper dbHelper = DatabaseHelper.getDatabaseHelper(context);
             dbHelper.insertOrReplaceKeyValue(DEVICE_ID_KEY, deviceId);
 
-            // remove device id from sharedPrefs so that this upgrade occurs only once
+            // remove deviceId from sharedPrefs so that this upgrade occurs only once
             preferences.edit().remove(Constants.PREFKEY_DEVICE_ID).apply();
         }