+++ src/com/amplitude/unity/plugins/AmplitudePlugin.java
@@ -4,11 +4,22 @@ import android.app.Application;
 import android.content.Context;
 
 import com.amplitude.api.Amplitude;
+import com.amplitude.api.Identify;
 
 import org.json.JSONException;
 import org.json.JSONObject;
 
 public class AmplitudePlugin {
+
+    public static JSONObject ToJSONObject(String jsonString) {
+        JSONObject properties = null;
+        try {
+            properties = new JSONObject(jsonString);
+        } catch (JSONException e) {
+            e.printStackTrace();
+        }
+        return properties;
+    }
 
     public static void init(Context context, String apiKey) {
         Amplitude.getInstance().initialize(context, apiKey);
@@ -33,25 +44,11 @@ public class AmplitudePlugin {
     }
 
     public static void logEvent(String event, String jsonProperties) {
-        JSONObject properties = null;
-        try {
-            properties = new JSONObject(jsonProperties);
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        Amplitude.getInstance().logEvent(event, properties);
+        Amplitude.getInstance().logEvent(event, ToJSONObject(jsonProperties));
     }
 
     public static void logEvent(String event, String jsonProperties, boolean outOfSession) {
-        JSONObject properties = null;
-        try {
-            properties = new JSONObject(jsonProperties);
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        Amplitude.getInstance().logEvent(event, properties, outOfSession);
+        Amplitude.getInstance().logEvent(event, ToJSONObject(jsonProperties), outOfSession);
     }
 
     public static void setUserId(String userId) {
@@ -63,14 +60,7 @@ public class AmplitudePlugin {
     }
 
     public static void setUserProperties(String jsonProperties) {
-        JSONObject properties = null;
-        try {
-            properties = new JSONObject(jsonProperties);
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-
-        Amplitude.getInstance().setUserProperties(properties);
+        Amplitude.getInstance().setUserProperties(ToJSONObject(jsonProperties));
     }
 
     public static void logRevenue(double amount) {
@@ -91,5 +81,231 @@ public class AmplitudePlugin {
 
     public static void trackSessionEvents(boolean enabled) {
         Amplitude.getInstance().trackSessionEvents(enabled);
+    }
+
+    // User Property Operations
+
+    // clear user properties
+    public static void clearUserProperties() {
+        Amplitude.getInstance().clearUserProperties();
+    }
+
+    // unset user property
+    public static void unsetUserProperty(String property) {
+        Amplitude.getInstance().identify(new Identify().unset(property));
+    }
+
+    // setOnce user property
+    public static void setOnceUserProperty(String property, boolean value) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, value));
+    }
+
+    public static void setOnceUserProperty(String property, double value) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, value));
+    }
+
+    public static void setOnceUserProperty(String property, float value) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, value));
+    }
+
+    public static void setOnceUserProperty(String property, int value) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, value));
+    }
+
+    public static void setOnceUserProperty(String property, long value) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, value));
+    }
+
+    public static void setOnceUserProperty(String property, String value) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, value));
+    }
+
+    public static void setOnceUserPropertyDict(String property, String values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, ToJSONObject(values)));
+    }
+
+    public static void setOnceUserPropertyList(String property, String values) {
+        JSONObject properties = ToJSONObject(values);
+        if (properties == null) {
+            return;
+        }
+        Amplitude.getInstance().identify(new Identify().setOnce(
+            property, properties.optJSONArray("list")
+        ));
+    }
+
+    public static void setOnceUserProperty(String property, boolean[] values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, values));
+    }
+
+    public static void setOnceUserProperty(String property, double[] values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, values));
+    }
+
+    public static void setOnceUserProperty(String property, float[] values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, values));
+    }
+
+    public static void setOnceUserProperty(String property, int[] values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, values));
+    }
+
+    public static void setOnceUserProperty(String property, long[] values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, values));
+    }
+
+    public static void setOnceUserProperty(String property, String[] values) {
+        Amplitude.getInstance().identify(new Identify().setOnce(property, values));
+    }
+
+    // set user property
+    public static void setUserProperty(String property, boolean value) {
+        Amplitude.getInstance().identify(new Identify().set(property, value));
+    }
+
+    public static void setUserProperty(String property, double value) {
+        Amplitude.getInstance().identify(new Identify().set(property, value));
+    }
+
+    public static void setUserProperty(String property, float value) {
+        Amplitude.getInstance().identify(new Identify().set(property, value));
+    }
+
+    public static void setUserProperty(String property, int value) {
+        Amplitude.getInstance().identify(new Identify().set(property, value));
+    }
+
+    public static void setUserProperty(String property, long value) {
+        Amplitude.getInstance().identify(new Identify().set(property, value));
+    }
+
+    public static void setUserProperty(String property, String value) {
+        Amplitude.getInstance().identify(new Identify().set(property, value));
+    }
+
+    public static void setUserPropertyDict(String property, String values) {
+        Amplitude.getInstance().identify(new Identify().set(property, ToJSONObject(values)));
+    }
+
+    public static void setUserPropertyList(String property, String values) {
+        JSONObject properties = ToJSONObject(values);
+        if (properties == null) {
+            return;
+        }
+        Amplitude.getInstance().identify(new Identify().set(
+                property, properties.optJSONArray("list")
+        ));
+    }
+
+    public static void setUserProperty(String property, boolean[] values) {
+        Amplitude.getInstance().identify(new Identify().set(property, values));
+    }
+
+    public static void setUserProperty(String property, double[] values) {
+        Amplitude.getInstance().identify(new Identify().set(property, values));
+    }
+
+    public static void setUserProperty(String property, float[] values) {
+        Amplitude.getInstance().identify(new Identify().set(property, values));
+    }
+
+    public static void setUserProperty(String property, int[] values) {
+        Amplitude.getInstance().identify(new Identify().set(property, values));
+    }
+
+    public static void setUserProperty(String property, long[] values) {
+        Amplitude.getInstance().identify(new Identify().set(property, values));
+    }
+
+    public static void setUserProperty(String property, String[] values) {
+        Amplitude.getInstance().identify(new Identify().set(property, values));
+    }
+
+    // add
+    public static void addUserProperty(String property, double value) {
+        Amplitude.getInstance().identify(new Identify().add(property, value));
+    }
+
+    public static void addUserProperty(String property, float value) {
+        Amplitude.getInstance().identify(new Identify().add(property, value));
+    }
+
+    public static void addUserProperty(String property, int value) {
+        Amplitude.getInstance().identify(new Identify().add(property, value));
+    }
+
+    public static void addUserProperty(String property, long value) {
+        Amplitude.getInstance().identify(new Identify().add(property, value));
+    }
+
+    public static void addUserProperty(String property, String value) {
+        Amplitude.getInstance().identify(new Identify().add(property, value));
+    }
+
+    public static void addUserPropertyDict(String property, String values) {
+        Amplitude.getInstance().identify(new Identify().add(property, ToJSONObject(values)));
+    }
+
+    // append user property
+    public static void appendUserProperty(String property, boolean value) {
+        Amplitude.getInstance().identify(new Identify().append(property, value));
+    }
+
+    public static void appendUserProperty(String property, double value) {
+        Amplitude.getInstance().identify(new Identify().append(property, value));
+    }
+
+    public static void appendUserProperty(String property, float value) {
+        Amplitude.getInstance().identify(new Identify().append(property, value));
+    }
+
+    public static void appendUserProperty(String property, int value) {
+        Amplitude.getInstance().identify(new Identify().append(property, value));
+    }
+
+    public static void appendUserProperty(String property, long value) {
+        Amplitude.getInstance().identify(new Identify().append(property, value));
+    }
+
+    public static void appendUserProperty(String property, String value) {
+        Amplitude.getInstance().identify(new Identify().append(property, value));
+    }
+
+    public static void appendUserPropertyDict(String property, String values) {
+        Amplitude.getInstance().identify(new Identify().append(property, ToJSONObject(values)));
+    }
+
+    public static void appendUserPropertyList(String property, String values) {
+        JSONObject properties = ToJSONObject(values);
+        if (properties == null) {
+            return;
+        }
+        Amplitude.getInstance().identify(new Identify().append(
+                property, properties.optJSONArray("list")
+        ));
+    }
+
+    public static void appendUserProperty(String property, boolean[] values) {
+        Amplitude.getInstance().identify(new Identify().append(property, values));
+    }
+
+    public static void appendUserProperty(String property, double[] values) {
+        Amplitude.getInstance().identify(new Identify().append(property, values));
+    }
+
+    public static void appendUserProperty(String property, float[] values) {
+        Amplitude.getInstance().identify(new Identify().append(property, values));
+    }
+
+    public static void appendUserProperty(String property, int[] values) {
+        Amplitude.getInstance().identify(new Identify().append(property, values));
+    }
+
+    public static void appendUserProperty(String property, long[] values) {
+        Amplitude.getInstance().identify(new Identify().append(property, values));
+    }
+
+    public static void appendUserProperty(String property, String[] values) {
+        Amplitude.getInstance().identify(new Identify().append(property, values));
     }
 }