+++ test/com/amplitude/api/InitializeTest.java
@@ -2,14 +2,14 @@ package com.amplitude.api;
 
 import android.content.Context;
 
-import com.squareup.okhttp.mockwebserver.RecordedRequest;
-
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
+
+import okhttp3.mockwebserver.RecordedRequest;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -20,7 +20,6 @@ public class InitializeTest extends BaseTest {
 
     @Before
     public void setUp() throws Exception {
-        amplitude = Amplitude.getInstance();
         super.setUp();
     }