+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Update to OKHttp v3.0.1.
+
 ## 2.5.1 (March 14, 2016)
 
 * Fix bug where updateServer sets the wrong batchLimit when limit is false.
@@ -7,7 +9,7 @@
 ## 2.5.0 (January 15, 2016)
 
 * Add ability to clear all user properties.
-* Check that SDK is initialized when user calls enableForegroundTracking, identify, setUserProperties
+* Check that SDK is initialized when user calls enableForegroundTracking, identify, setUserProperties.
 
 ## 2.4.0 (December 15, 2015)